refactor(config): split nimbus-toml.ts by responsibility + dedup loaders
nimbus-toml.ts was a 1003-line monolith mixing per-section config loaders
with the DORA/CI service-config machinery. Split into three cohesive files:
- toml-primitives.ts (52): shared scalar/line parsers (stripComment,
  isTableHeader, splitKeyValue, parseString, parseIntDec, parseStringArray).
- service-config-toml.ts (203): the DORA + [ci.service.*] ServiceConfig
  parsing/materialization.
- nimbus-toml.ts (729): per-section types/defaults/parsers/loaders; barrel
  re-exports service-config-toml for back-compat (external importers
  unchanged).

Also dedup the ~14 near-identical section loaders behind one generic
loadTomlSection<T>(path, fallback, parse), preserving each load function's
exact variant (merge-partial vs full-parse). loadNimbusPagerdutyFromPath
kept verbatim (it emits distinct stderr diagnostics).

Imports are strictly acyclic (primitives ← both; service-config ← nimbus-toml
only). Behavior-preserving: gateway tsc clean, 180 config+metrics tests pass,
audit:invariants exit 0 (I6 LAN bind default intact), biome clean. Per-file
coverage ≥80% (nimbus-toml 82.1%, service-config 97.5%, primitives 94.4%).
